// src/app/models/adjunto.ts
export class Adjunto {
  uidCarga?: string;
  uidRegistro?: number;
  uidAdjunto?: number;
  tipoOrigen?: string;
  tipo?: string;
  ruta?: string;
  nombre?: string;
  extension?: string;
  estado?: string;
  usuarioCreacion?: string;
  usuarioModificacion?: string;
  fechaCreacion?: Date;
  fechaModificacion?: Date;
  fechaCarga?: string
}